requirements.txt
pytest==6.2.5
appium-python-client==2.1.0
requests==2.26.0
allure-pytest==2.9.45